# sqlite: don't cache sqlite column types between rows