If we had a NULL in the first row, we were assuming all rows thereafter had a NULL in that column.

Further, SQLite lets you have different types per row even within the same column, so caching the column type was kinda wrong in general anyway.

Presumably this was cached previously for performance reasons, but the handful of commits prior to this one made the cgo calls alloc-free, so this is much cheaper than it used to be.

                             │   before    │                after                │
                             │   sec/op    │   sec/op     vs base                │
    WALHookAndExec-8           2.316µ ± 2%   2.327µ ± 1%        ~ (p=0.254 n=10)
    Persist-8                  1.980µ ± 1%   2.037µ ± 1%   +2.85% (p=0.001 n=10)
    QueryRows100MixedTypes-8   36.44µ ± 2%   46.14µ ± 1%  +26.64% (p=0.000 n=10)
    EmptyExec-8                357.0n ± 2%   356.6n ± 1%        ~ (p=0.565 n=10)
    BeginTxNoop-8              9.178µ ± 0%   9.204µ ± 0%        ~ (p=0.239 n=10)
    geomean                    3.529µ        3.726µ        +5.56%

                             │    before    │                 after                 │
                             │     B/op     │     B/op      vs base                 │
    WALHookAndExec-8             48.00 ± 0%     48.00 ± 0%       ~ (p=1.000 n=10) ¹
    Persist-8                    456.0 ± 0%     424.0 ± 0%  -7.02% (p=0.000 n=10)
    QueryRows100MixedTypes-8   2.719Ki ± 0%   2.688Ki ± 0%  -1.15% (p=0.000 n=10)
    EmptyExec-8                  32.00 ± 0%     32.00 ± 0%       ~ (p=1.000 n=10) ¹
    BeginTxNoop-8                624.0 ± 0%     624.0 ± 0%       ~ (p=1.000 n=10)
    geomean                      261.2          256.9       -1.67%
    ¹ all samples are equal

                             │   before   │                after                │
                             │ allocs/op  │ allocs/op   vs base                 │
    WALHookAndExec-8           2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹
    Persist-8                  11.00 ± 0%   11.00 ± 0%       ~ (p=1.000 n=10) ¹
    QueryRows100MixedTypes-8   107.0 ± 0%   107.0 ± 0%       ~ (p=1.000 n=10) ¹
    EmptyExec-8                1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
    BeginTxNoop-8              13.00 ± 0%   13.00 ± 0%       ~ (p=1.000 n=10) ¹
    geomean                    7.891        7.891       +0.00%
    ¹ all samples are equal

Fixes #32