@raggi 2023-04-09
> 🤯 TIL that SQLite allows the type to vary across rows. Is that only within affinities, or would it (say) allow TEXT and INTEGER to mix?

Switching from default to SQLite strict tables is a bit like switching to innodb in MySQL. https://www.sqlite.org/stricttables.html

Nullability would still be a thing though.