@bradfitz 2023-04-09
> 🤯 TIL that SQLite allows the type to vary across rows. Is that only within affinities, or would it (say) allow TEXT and INTEGER to mix?

Anything. Look at the `changing-types-within-a-column-between-rows` test case that I'd added which shows changing between string, integer, and null.