@creachadair 2023-04-09
> > 🤯 TIL that SQLite allows the type to vary across rows. Is that only within affinities, or would it (say) allow TEXT and INTEGER to mix?
> 
> Anything. Look at the `changing-types-within-a-column-between-rows` test case that I'd added which shows changing between string, integer, and null.

INTEGER/TEXT isn't surprising—the docs promise it will apply standard-ish coercions, which empirically includes parsing integer-shaped strings and rendering integers into strings. The fact that it will also let you jam floats (say) in there was the surprising part to me 😭