@bradfitz 2023-04-10
BTW, the benchmark numbers here may look bad but I have a follow-up change which does all the column_type lookups in the same cgo call as the sqlite3_step call. That actually makes it ~faster (barely), comparing from before this commit to after the two commits:

```
                         │   before    │               after2               │
                         │   sec/op    │   sec/op     vs base               │
WALHookAndExec-8           2.316µ ± 2%   2.297µ ± 1%       ~ (p=0.109 n=10)
Persist-8                  1.980µ ± 1%   1.956µ ± 0%  -1.26% (p=0.000 n=10)
QueryRows100MixedTypes-8   36.44µ ± 2%   36.47µ ± 1%       ~ (p=0.289 n=10)
EmptyExec-8                357.0n ± 2%   355.1n ± 1%       ~ (p=0.697 n=10)
BeginTxNoop-8              9.178µ ± 0%   9.178µ ± 2%       ~ (p=0.684 n=10)
geomean                    3.529µ        3.512µ       -0.50%

                         │    before    │                after2                 │
                         │     B/op     │     B/op      vs base                 │
WALHookAndExec-8             48.00 ± 0%     48.00 ± 0%       ~ (p=1.000 n=10) ¹
Persist-8                    456.0 ± 0%     440.0 ± 0%  -3.51% (p=0.000 n=10)
QueryRows100MixedTypes-8   2.719Ki ± 0%   2.703Ki ± 0%  -0.57% (p=0.000 n=10)
EmptyExec-8                  32.00 ± 0%     32.00 ± 0%       ~ (p=1.000 n=10) ¹
BeginTxNoop-8                624.0 ± 0%     624.0 ± 0%       ~ (p=1.000 n=10)
geomean                      261.2          259.1       -0.83%
¹ all samples are equal

                         │   before   │               after2                │
                         │ allocs/op  │ allocs/op   vs base                 │
WALHookAndExec-8           2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹
Persist-8                  11.00 ± 0%   11.00 ± 0%       ~ (p=1.000 n=10) ¹
QueryRows100MixedTypes-8   107.0 ± 0%   107.0 ± 0%       ~ (p=1.000 n=10) ¹
EmptyExec-8                1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
BeginTxNoop-8              13.00 ± 0%   13.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                    7.891        7.891       +0.00%
¹ all samples are equal
```